Add new UI/UX interaction on add variables

+++ src/app/portlets/content-types/fields/content-type-fields-variables/content-type-fields-variables.component.ts
@@ -3,7 +3,6 @@ import { DotMessageService } from '../../../../api/services/dot-messages-service
 import { FieldVariablesService, FieldVariableParams } from '../service/';
 import { DotHttpErrorManagerService } from '../../../../api/services/dot-http-error-manager/dot-http-error-manager.service';
 import { ResponseView } from 'dotcms-js';
-import { Column } from 'primeng/primeng';
 import { take, takeUntil } from 'rxjs/operators';
 import { Subject } from 'rxjs/internal/Subject';
 
@@ -26,7 +25,6 @@ export class ContentTypeFieldsVariablesComponent implements OnInit, OnChanges, O
 
     fieldVariables: FieldVariable[] = [];
     messages: { [key: string]: string } = {};
-    canSaveFields: boolean[] | null[] = [];
     private destroy$: Subject<boolean> = new Subject<boolean>();
 
     constructor(
@@ -41,7 +39,9 @@ export class ContentTypeFieldsVariablesComponent implements OnInit, OnChanges, O
                 'contenttypes.field.variables.key_header.label',
                 'contenttypes.field.variables.value_header.label',
                 'contenttypes.field.variables.actions_header.label',
-                'contenttypes.field.variables.value_no_rows.label'
+                'contenttypes.field.variables.value_no_rows.label',
+                'contenttypes.action.save',
+                'contenttypes.action.cancel'
             ])
             .pipe(take(1))
             .subscribe((messages: { [key: string]: string }) => {
@@ -66,23 +66,51 @@ export class ContentTypeFieldsVariablesComponent implements OnInit, OnChanges, O
      * @param {FieldVariable} variable
      * @memberof ContentTypeFieldsVariablesComponent
      */
-    deleteVariable(fieldIndex: number): void {
-        const params: FieldVariableParams = {
-            contentTypeId: this.field.contentTypeId,
-            fieldId: this.field.fieldId,
-            variable: this.fieldVariables[fieldIndex]
-        };
-        this.fieldVariablesService.delete(params).pipe(take(1)).subscribe(
-            () => {
-                this.fieldVariables = this.fieldVariables.filter(
-                    (_item: FieldVariable, index: number) => index !== fieldIndex
-                );
-                this.canSaveFields.splice(fieldIndex, 1);
-            },
-            (err: ResponseView) => {
-                this.dotHttpErrorManagerService.handle(err).pipe(take(1)).subscribe();
-            }
-        );
+    deleteVariable(fieldVariable: FieldVariable, fieldIndex: number): void {
+        if (fieldVariable.key !== '') {
+            const params: FieldVariableParams = {
+                contentTypeId: this.field.contentTypeId,
+                fieldId: this.field.fieldId,
+                variable: this.fieldVariables[fieldIndex]
+            };
+            this.fieldVariablesService.delete(params).pipe(take(1)).subscribe(
+                () => {
+                    this.fieldVariables = this.fieldVariables.filter(
+                        (_item: FieldVariable, index: number) => index !== fieldIndex
+                    );
+                },
+                (err: ResponseView) => {
+                    this.dotHttpErrorManagerService.handle(err).pipe(take(1)).subscribe();
+                }
+            );
+        } else {
+            this.fieldVariables = this.fieldVariables.filter(
+                (_item: FieldVariable, index: number) => index !== fieldIndex
+            );
+        }
+    }
+
+    /**
+     * Handle Edit button event
+     * @param {Number} fieldIndex
+     * @memberof ContentTypeFieldsVariablesComponent
+     */
+    editVariableAction(fieldIndex: number): void {
+        const key = document.getElementById(`content-type-fields__variables-key-${fieldIndex}`);
+        key.click();
+        this.showEditVariableAction(fieldIndex);
+    }
+
+    /**
+     * Handle Cancel button event
+     * @param {Number} fieldIndex
+     * @memberof ContentTypeFieldsVariablesComponent
+     */
+    cancelVariableAction(fieldIndex: number): void {
+        const editMenu = document.getElementById(`content-type-fields__variables-actions-edit-${fieldIndex}`);
+        const mainMenu = document.getElementById(`content-type-fields__variables-actions-main-${fieldIndex}`);
+        mainMenu.style.display = 'block';
+        editMenu.style.display = 'none';
     }
 
     /**
@@ -96,36 +124,63 @@ export class ContentTypeFieldsVariablesComponent implements OnInit, OnChanges, O
             fieldId: this.field.fieldId,
             variable: variable
         };
-        this.fieldVariablesService.save(params).pipe(take(1)).subscribe(
-            (savedVariable: FieldVariable) => {
-                if (typeof variableIndex !== 'undefined') {
-                    this.fieldVariables = this.updateVariableCollection(
-                        savedVariable,
-                        variableIndex
-                    );
-                } else {
-                    this.fieldVariables = [].concat(savedVariable, this.fieldVariables);
-                    this.canSaveFields = [].concat(null, this.canSaveFields);
+        if (typeof variableIndex !== 'undefined') {
+            this.fieldVariablesService.save(params).pipe(take(1)).subscribe(
+                (savedVariable: FieldVariable) => {
+                    if (typeof variableIndex !== 'undefined') {
+                        this.fieldVariables = this.updateVariableCollection(
+                            savedVariable,
+                            variableIndex
+                        );
+                    } else {
+                        this.fieldVariables = [].concat(savedVariable, this.fieldVariables);
+                    }
+                },
+                (err: ResponseView) => {
+                    this.dotHttpErrorManagerService.handle(err).pipe(take(1)).subscribe();
                 }
-            },
-            (err: ResponseView) => {
-                this.dotHttpErrorManagerService.handle(err).pipe(take(1)).subscribe();
-            }
-        );
+            );
+        } else {
+            const emptyVariable: FieldVariable = {
+                key: '',
+                value: ''
+            };
+            this.fieldVariables = [].concat(emptyVariable, this.fieldVariables);
+        }
     }
 
     /**
-     * Holds status if a existing variable meets conditions to be updated
-     * @param {any} event
+     * Handle Add new variable button event
      * @memberof ContentTypeFieldsVariablesComponent
      */
-    updateSaveButtons(event: {
-        column: Column;
-        data: { [key: string]: string };
-        index: number;
-    }): void {
-        this.canSaveFields[event.index] =
-            event.data.key.length > 0 && event.data.value.length > 0 ? null : true;
+    addVariable(): void {
+        this.saveVariable(null);
+        setTimeout(() => {
+            this.editVariableAction(0);
+        });
+    }
+
+    /**
+     * Handle onFocus event on Variable's input fields
+     * @param {FieldVariable} fieldVariable
+     * @param {Number} index
+     * @memberof ContentTypeFieldsVariablesComponent
+     */
+    editFieldInit(fieldVariable: FieldVariable, index: number): void {
+        const disabled = fieldVariable.key === '' || fieldVariable.value === '' ? true : false;
+        this.showEditVariableAction(index, disabled);
+    }
+
+    private showEditVariableAction(fieldIndex: number, disabled?: boolean): void {
+        const editMenu = document.getElementById(`content-type-fields__variables-actions-edit-${fieldIndex}`);
+        const mainMenu = document.getElementById(`content-type-fields__variables-actions-main-${fieldIndex}`);
+        editMenu.style.display = 'block';
+        mainMenu.style.display = 'none';
+        if (disabled) {
+            editMenu.getElementsByTagName('button')[1].disabled = true;
+        } else {
+            editMenu.getElementsByTagName('button')[1].disabled = false;
+        }
     }
 
     private initTableData(): void {
